Fix report files size

+++ services/mass_storage_service.c
@@ -1067,7 +1067,7 @@ SAT_returnState get_fs_stat(uint8_t *buf, uint16_t *size) {
 
         } else if(sid <= EVENT_LOG) {
 
-            uint16_t files_num = head - tail + 1;
+            uint16_t files_num = (uint16_t)(head - tail + 1);
 
             cnv16_8(files_num, &buf[(*size)]);
             *size += sizeof(uint16_t);
